requirements.txt
runpod==1.5.0
pyannote.audio
torch>=2.0.0
torchaudio>=2.0.0 